PIM-8665 Uses a command to rotate SSO log (#7385)
* PIM-8665 Uses command for SSO log rotation instead of doing it in the close() of the handler

* PIM-8665 Tests SSO log rotate command

* PIM-8665 Adds cron to rotate logs on SaaS

* PIM-8665 Updates migration

* PIM-8665 Fixes SSO test with no rotate

* Removes obsolete migration

* PIM-8665 Rollbacks deleted migration file

* PIM-8665 Fixes phpSpec version

+++ src/Akeneo/Platform/Bundle/InstallerBundle/Command/EnterpriseDatabaseCommand.php
@@ -92,7 +92,7 @@ SQL
             CREATE TABLE pimee_sso_log (
               time DATETIME,
               channel VARCHAR(255),
-              level TINYINT,
+              level SMALLINT,
               message TEXT,
               INDEX(time)
             ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
